requirements.txt
# HDRi 360 Studio - OpenCV Panorama Processing Server Requirements
# Compatible with Python 3.7+

# Web framework (Python 3.7 compatible versions)
Flask==2.2.5
Werkzeug==2.2.3

# Computer vision and image processing (Python 3.7 compatible)
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
numpy==1.21.6
Pillow==9.5.0

# HTTP requests for server communication
requests==2.28.2

# Additional utilities
python-dateutil==2.8.2